EDSC-4303: Refactor map tests
