+++ static/src/js/components/Map/ShapefileLayerExtended.jsx
@@ -24,6 +24,7 @@ export class ShapefileLayerExtended extends L.Layer {
     this.isProjectPage = props.isProjectPage
     this.onToggleTooManyPointsModal = props.onToggleTooManyPointsModal
     this.onChangeProjection = props.onChangeProjection
+    this.projection = props.projection
 
     this.options = {
       selection: L.extend({}, defaultOptions.selection, colorOptions)
@@ -190,6 +191,7 @@ export class ShapefileLayerExtended extends L.Layer {
     const layersToSelect = []
     let allLatsArctic = false
     let allLatsAntarctic = false
+
     // eslint-disable-next-line new-cap
     const jsonLayer = new L.geoJson(response, {
       className: 'geojson-svg',
@@ -239,10 +241,24 @@ export class ShapefileLayerExtended extends L.Layer {
       }
     })
 
+    if (allLatsArctic && this.projection !== projections.arctic) {
+      // Change projection to arctic
+      this.onChangeProjection(projections.arctic)
+
+      return
+    }
+
+    if (allLatsAntarctic && this.projection !== projections.antarctic) {
+      // Change projection to arctic
+      this.onChangeProjection(projections.antarctic)
+
+      return
+    }
+
     if (!this.isProjectPage) jsonLayer.on('click', this.clickLayer)
     this.jsonLayer = jsonLayer
     this.jsonLayer.addTo(this.map)
-    if (panMap) this.map.flyToBounds(jsonLayer.getBounds())
+    if (panMap) this.map.flyToBounds(jsonLayer.getBounds(), { animate: false })
     this.onMetricsMap('Added Shapefile')
 
     const children = jsonLayer.getLayers()
@@ -262,16 +278,6 @@ export class ShapefileLayerExtended extends L.Layer {
     const fileHash = forge.md.md5.create()
     fileHash.update(JSON.stringify(response))
     this.fileHash = fileHash.digest().toHex()
-
-    if (allLatsArctic) {
-      // Change projection to arctic
-      this.onChangeProjection(projections.arctic)
-    }
-
-    if (allLatsAntarctic) {
-      // Change projection to arctic
-      this.onChangeProjection(projections.antarctic)
-    }
   }
 
   // Fired when a shapefile layer, or selected shape layer is clicked
