+++ tests/e2e/collection_list/collection_list.spec.js
@@ -5,6 +5,7 @@ import singleCollection from './__mocks__/single_collection.json'
 test.describe('Collection List Behavior', () => {
   test.beforeEach(async ({ page }) => {
     await page.route('**/*.{png,jpg,jpeg}', (route) => route.abort())
+    await page.route('**/scale/**', (route) => route.abort())
     await page.route(/collections.json/, async (route) => {
       await route.fulfill({
         json: singleCollection.body,
