+++ tests/e2e/map/__mocks__/bounding_box_collections_edited.body.json
