+++ tests/e2e/map/map_granules.spec.js-snapshots/chromium/colormap-2-screenshot.png
