+++ tests/e2e/map/map_granules.spec.js-snapshots/firefox/colormap-2-screenshot.png
