+++ tests/e2e/map/map_granules.spec.js-snapshots/firefox/colormap-screenshot.png
