+++ tests/e2e/map/map_granules.spec.js-snapshots/webkit/colormap-2-screenshot.png
