+++ tests/e2e/paths/search/search.spec.js
@@ -81,6 +81,7 @@ const testResultsSize = async (page, cmrHits) => {
 test.describe('Path /search', () => {
   test.beforeEach(async ({ page }) => {
     await page.route('**/*.{png,jpg,jpeg}', (route) => route.abort())
+    await page.route('**/scale/**', (route) => route.abort())
   })
 
   test.describe('When the path is loaded without any url params', () => {
@@ -222,7 +223,7 @@ test.describe('Path /search', () => {
     })
   })
 
-  test.describe.skip('When the path is loaded with a spatial query', () => {
+  test.describe('When the path is loaded with a spatial query', () => {
     test.describe('When the spatial query is a point', () => {
       test('loads with the spatial query applied', async ({ page }) => {
         const cmrHits = 5079
@@ -255,8 +256,8 @@ test.describe('Path /search', () => {
 
         await expect(page.getByTestId('spatial-display_point')).toHaveValue('4.33676,65.44171')
 
-        // Test leaflet has drawn the shape correctly
-        await expect(page.locator('.leaflet-marker-pane img')).toHaveAttribute('style', 'margin-left: -12px; margin-top: -41px; width: 25px; height: 41px; transform: translate3d(1165px, 383px, 0px); z-index: 383;')
+        // Test leaflet has drawn the point
+        await expect(await page.getByRole('button', { name: 'Marker' }).all()).toHaveLength(1)
       })
     })
 
@@ -292,8 +293,9 @@ test.describe('Path /search', () => {
 
         await expect(page.getByTestId('spatial-display_polygon')).toHaveText('3 Points')
 
-        // Test leaflet has drawn the shape correctly
-        await expect(page.locator('.leaflet-interactive')).toHaveAttribute('d', 'M1161 404L1122 480L1257 499L1161 404z')
+        // Test leaflet has drawn the shape
+        await expect(await page.locator('g path').first()).toBeVisible()
+        await expect(await page.locator('g path').all()).toHaveLength(1)
       })
     })
 
@@ -330,8 +332,9 @@ test.describe('Path /search', () => {
         await expect(page.getByTestId('spatial-display_circle-center')).toHaveValue('2.22154,62.18209')
         await expect(page.getByTestId('spatial-display_circle-radius')).toHaveValue('100000')
 
-        // Test leaflet has drawn the shape correctly
-        await expect(page.locator('.leaflet-interactive')).toHaveAttribute('d', 'M1136.1837511111112,398.20238222222224a6,6 0 1,0 12,0 a6,6 0 1,0 -12,0 ')
+        // Test leaflet has drawn the shape
+        await expect(await page.locator('g path').first()).toBeVisible()
+        await expect(await page.locator('g path').all()).toHaveLength(1)
       })
     })
 
@@ -368,8 +371,9 @@ test.describe('Path /search', () => {
         await expect(page.getByTestId('spatial-display_southwest-point')).toHaveValue('0.99949,5.02679')
         await expect(page.getByTestId('spatial-display_northeast-point')).toHaveValue('26.17555,32.8678')
 
-        // Test leaflet has drawn the shape correctly
-        await expect(page.locator('.leaflet-interactive')).toHaveAttribute('d', 'M736 407L736 228L934 228L934 407L736 407z')
+        // Test leaflet has drawn the shape
+        await expect(await page.locator('g path').first()).toBeVisible()
+        await expect(await page.locator('g path').all()).toHaveLength(1)
       })
     })
 
@@ -442,8 +446,9 @@ test.describe('Path /search', () => {
         await expect(page.getByTestId('spatial-display_shapefile-name')).toHaveText('test.geojson')
         await expect(page.getByTestId('filter-stack-item__hint')).toHaveText('1 shape selected')
 
-        // Test leaflet has drawn the shape correctly
-        await expect(page.locator('.leaflet-interactive').first()).toHaveAttribute('d', 'M1161 404L1122 480L1257 499L1161 404z')
+        // Test leaflet has drawn the shape
+        await expect(await page.locator('g path').first()).toBeVisible()
+        await expect(await page.locator('g path').all()).toHaveLength(2)
       })
     })
   })
