+++ tests/e2e/tour/tour.spec.js
@@ -12,6 +12,8 @@ const expectWithinMargin = async (actual, expected, margin) => {
 test.describe('Joyride Tour Navigation', () => {
   test.beforeEach(async ({ page }) => {
     await page.route('**/*.{png,jpg,jpeg}', (route) => route.abort())
+    await page.route('**/scale/**', (route) => route.abort())
+
     await page.route(/collections.json/, async (route) => {
       await route.fulfill({
         json: singleCollection.body,
